Texture
Window position can now be set on creation, ~0U puts the window at the center of the display1, applied separatelly for x and y.
Uncompressed TGA files are loaded onto ram and using a staging image, put on device memory.
A new DescriptorSetLayout handles the new CombinedImageSampler defined in the shaders; A new DescriptorSet created from that layout tells the shader what texture to use as diffuse.
Fragment shader now defines color from texture instead of using a predefined color.
A single stagingTexture is used for any number of textures, the only cap is the predefined extent2D of the stagingTexture, neither width of height can surpass it; If anything smaller is used, the proper size is blitted from the stagingTexture.